requirements.txt
requests==2.20.0
beautifulsoup4==4.9.3
PyExecJS==1.5.1
